Investigate current SdfPath validation for Unicode inputs
Added basic support for Unicode UTF-8 encoded prim and property names.
Separated out logic for validation of prim names vs. property names.
Added validation logic for Unicode XID_Start / XID_Continue character classes for property names.
Adjusted lexers and parsers to accept and interpret UTF-8 characters.
Added additional basic tests for UTF-8 prim paths and UTF-8 encoded usda file reading with Unicode prim names.
Added UnicodeDatabase.txt for reference on character classes.
Added new environment setting ARCH_UTF8_IDENTIFIERS that controls which logic is used to validate prim / property names.

Currently only validated on Windows.
Failed tests need discussion.

+++ pxr/usd/sdf/pathParser.h
@@ -60,7 +60,7 @@ int pathYyparse(Sdf_PathParserContext *context);
 int pathYylex_init(yyscan_t *yyscanner);
 int pathYylex_destroy(yyscan_t yyscanner);
 yy_buffer_state *pathYy_scan_string(const char* str, yyscan_t yyscanner);
-yy_buffer_state *pathYy_scan_bytes(const char* str, size_t len, 
+yy_buffer_state *pathYy_scan_bytes(const char* str, int len, 
                                    yyscan_t yyscanner);
 void pathYy_delete_buffer(yy_buffer_state *b, yyscan_t yyscanner);
